Fixed simaese_loss in model_specifications and train.py
changed siames to siamese_loss in specifications.json to match the name with model name used in train.py. Changed line 81 and 96 in train.py to recognize siamese_loss. Earlier siamese_loss was being accessed directly while it was a key in the diction "__loss_options__". Now it is being accessed through __loss_options__.